Fix a bug where calibration routines could potentially overwrite the early memory areas in EEPROM

+++ speeduino/storage.ino
@@ -652,6 +652,6 @@ void writeCalibration()
 // By having these in this file, it prevents other files from calling EEPROM functions directly. This is useful due to differences in the EEPROM libraries on different devces
 byte readLastBaro() { return EEPROM.read(EEPROM_LAST_BARO); }
 void storeLastBaro(byte newValue) { EEPROM.update(EEPROM_LAST_BARO, newValue); }
-void storeCalibrationValue(byte location, byte value) { EEPROM.update(location, value); } //This is essentially just an abstraction for EEPROM.update()
+void storeCalibrationValue(uint16_t location, byte value) { EEPROM.update(location, value); } //This is essentially just an abstraction for EEPROM.update()
 byte readEEPROMVersion() { return EEPROM.read(EEPROM_DATA_VERSION); }
 void storeEEPROMVersion(byte newVersion) { EEPROM.update(EEPROM_DATA_VERSION, newVersion); }
